Add more validations

+++ Assets/Scripts/JsonManager.cs
@@ -1,26 +1,25 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
 using UnityEngine;
-using UnityEngine.UIElements;
-using Random = UnityEngine.Random;
-using System.Runtime.Serialization;
 using System.IO;
-using System.Text;
 
-public static class JsonManager{
+public static class JsonManager {
     public static bool SaveToJson(string fileName, GameData data) {
-        //validar que exista
+        // No tira error el writeAll pero está bueno chequear que no sea read only de todas formas:
+        if (File.Exists(Application.dataPath + "/Scripts/" + fileName) && ((File.GetAttributes(Application.dataPath + "/Scripts/" + fileName) & FileAttributes.ReadOnly)) != 0)
+            return false;
+        
         string textJson = JsonUtility.ToJson(data);
         File.WriteAllText(Application.dataPath + "/Scripts/" + fileName, textJson);
         return true;
     }
 
-    public static bool ReadFromJson(string fileName, out GameData dataBuff){
-        //validar que exista
+    public static bool ReadFromJson(string fileName, out GameData dataBuff) {
+        if (!File.Exists(Application.dataPath + "/Scripts/" + fileName)) {
+            dataBuff = null;
+            return false;
+        }
+
         string textJson = File.ReadAllText(Application.dataPath + "/Scripts/" + fileName);
-        dataBuff= JsonUtility.FromJson<GameData>(textJson);
+        dataBuff = JsonUtility.FromJson<GameData>(textJson);
         return true;
     }
 }
